Support excluding specific file patterns in nested directories
For a file pattern such as "dist/**/!(*.test.js)", previously it would
add an entry of "!dist/**/!(*.test.js)/**" for the ignore-walker, which
results in paths such as "dist/lib/x.test.js" getting included in the
pack. The fix is to avoid appending the "/**" if the file ends in
"**/!(pattern)".