Never include .git folders in package root
Git folders are in the default ignore set.  However, occasionally git
can create a file named 'readme', or another of npm's 'always include'
filenames.

This roots the always-include filename set to the root with a /, and
also avoids even walking the root's .git folder.  (Git folders can still
be included explicitly deeper in the package, if they are included in a
package.json files list, or un-ignored in a .gitignore or .npmignore.)