Revert "Exclude common core dump file names by default"
This reverts commits 1cd776cb1689429ef6fdb49df3097d81b1304256 and
800f03dcf2dec360050c57fa665925182e88e6b7.

While it's a great idea to not publish core dumps to the npm registry,
this approach is too blunt, and has broken a lot of packages in the
process.

Attempts to restrict the core dump file handling to avoid excluding
folders ends up including anything in any folder named core, which is
_also_ a problem.

An optimal solution would look at the actual files along the walk, and
filter them out if they're actually core dumps, using a more reliable
indicator than the filename "core".