Avoid ELOOP error when encountering symlinks in node_modules
This avoids a catastrophic failure where a symlink to a package in a
node_modules *other* than the main project node_modules (ie, in a test
repo or workspace) will cause the link to be followed, on the assumption
that it's a bundled dep or else we wouldn't be looking at it.

However, we only filter out non-bundled deps in the *main* node_modules
folder hierarchy, not in a node_modules folder that appears in some
other place in the project.

H/T @ljharb